ticket:4521: (hack) workaround, 64->32
git-svn-id: http://source.icu-project.org/repos/icu/icu@27548 251d0590-4201-4cf1-90de-194747b24ca1

+++ branches/srl/fix4521/source/i18n/regeximp.h
@@ -279,7 +279,7 @@ enum {
 //  Match Engine State Stack Frame Layout.
 //
 struct REStackFrame {
-    int64_t            fInputIdx;        // Position of next character in the input string
+    int32_t            fInputIdx;        // Position of next character in the input string
     int32_t            fPatIdx;          // Position of next Op in the compiled pattern
     int32_t            fExtra[2];        // Extra state, for capture group start/ends
                                          //   atomic parentheses, repeat counts, etc.
